fix: check DELETE_MANY cmd in middlewares to reformat keys
