+++ tests/test_middleware.py
@@ -1,5 +1,5 @@
 from cashews import Cache
-from cashews.helpers import add_prefix, all_keys_lower, memory_limit
+from cashews.helpers import all_keys_lower, memory_limit
 
 
 async def test_all_keys_lower(cache: Cache, target):
@@ -37,38 +37,3 @@ async def test_memory_limit(cache: Cache, target):
 
     await cache.ping()
     target.ping.assert_called_once_with(message=b"PING")
-
-
-async def test_add_prefix(cache: Cache, target):
-    cache._add_backend(target, (add_prefix("prefix!"),))
-
-    await cache.get(key="key")
-    target.get.assert_called_once_with(key="prefix!key", default=None)
-
-    await cache.set(key="key", value="value")
-    target.set.assert_called_once_with(
-        key="prefix!key",
-        value="value",
-        exist=None,
-        expire=None,
-    )
-    await cache.ping()
-    target.ping.assert_called_once_with(message=b"PING")
-
-
-async def test_add_prefix_get_many(cache: Cache, target):
-    cache._add_backend(target, (add_prefix("prefix!"),))
-    await cache.get_many("key")
-    target.get_many.assert_called_once_with("prefix!key")
-
-
-async def test_add_prefix_set_many(cache: Cache, target):
-    cache._add_backend(target, (add_prefix("prefix!"),))
-    await cache.set_many({"key": "value"})
-    target.set_many.assert_called_once_with(pairs={"prefix!key": "value"}, expire=None)
-
-
-async def test_add_prefix_delete_match(cache: Cache, target):
-    cache._add_backend(target, (add_prefix("prefix!"),))
-    await cache.delete_match("key")
-    target.delete_match.assert_called_once_with(pattern="prefix!key")
